+++ spec/features/wars_spec.rb
@@ -11,7 +11,7 @@ describe 'List wars' do
 
   specify { expect(subject.scope).to eq(nil) }
 
-  specify { expect(subject.wars_ids.size).to eq(2000) }
+  specify { expect(subject.war_ids.size).to eq(2000) }
 
-  specify { expect(subject.wars_ids.first).to eq(615_639) }
+  specify { expect(subject.war_ids.first).to eq(615_639) }
 end